requirements.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
dnspython==2.6.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.15.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.1
fsspec==2024.6.0
future==1.0.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.1
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
proglog==0.1.10
pymongo==4.7.3
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
